requirements.txt
chromadb==1.0.13
faiss_cpu==1.11.0
numpy==2.3.1
openai==1.90.0
pandas==2.3.0
Requests==2.32.4
